backend/pyproject.toml
[tool.poetry]
name = "jwt-fastapi-backend"
version = "0.1.0"
description = "FastAPI JWT example"
authors = ["copilot <copilot@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.111.0"
uvicorn = {extras = ["standard"], version = "0.30.1"}
PyJWT = "2.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
